split structs into separate modules
