+++ src/bonjour/mod.rs
@@ -1,22 +1,16 @@
-use std::path::{Path, PathBuf};
-
-//use crate::bonjour::differences::{PackageDataDifferences, AddedPackages, RetainedPackages, MergedPackageData};
+use crate::bonjour::changed_manifest_packages::ChangedManifestPackages;
+use crate::bonjour::installed_manifest_packages::InstalledManifestPackages;
 use crate::bonjour::lockfile::{LockfileData, LockfileResult, LockfileSource};
 use crate::bonjour::manifest::{ManifestData, ManifestResult, ManifestSource};
-use crate::cfg_toml::lock::lockfile_command::LockfileCommand;
-use crate::cfg_toml::lock::lockfile_module::LockfileModule;
-use graphql_client::*;
+use crate::bonjour::resolved_manifest_packages::ResolvedManifestPackages;
+use std::borrow::Cow;
+use std::path::Path;
 
+pub mod changed_manifest_packages;
+pub mod installed_manifest_packages;
 pub mod lockfile;
 pub mod manifest;
-
-#[derive(GraphQLQuery)]
-#[graphql(
-    schema_path = "graphql/schema.graphql",
-    query_path = "graphql/queries/get_packages.graphql",
-    response_derives = "Debug"
-)]
-struct GetPackagesQuery;
+pub mod resolved_manifest_packages;
 
 #[derive(Clone, Debug, Fail)]
 pub enum BonjourError {
@@ -56,192 +50,6 @@ impl<'a> PackageKey<'a> {
     }
 }
 
-#[derive(Clone, Debug)]
-pub struct LockfilePackage<'a> {
-    pub modules: Vec<LockfileModule<'a>>,
-    pub commands: Vec<LockfileCommand<'a>>,
-}
-
-use crate::cfg_toml::manifest::Manifest;
-use crate::graphql::execute_query;
-use crate::util::{
-    create_package_dir, fully_qualified_package_display_name, get_package_namespace_and_name,
-};
-use flate2::read::GzDecoder;
-use std::borrow::Cow;
-use std::collections::hash_set::HashSet;
-use std::fs::OpenOptions;
-use std::io;
-use std::io::SeekFrom;
-use tar::Archive;
-
-#[derive(Clone, Debug)]
-pub struct ResolvedManifestPackages<'a> {
-    pub packages: Vec<(WapmPackageKey<'a>, String)>,
-}
-
-impl<'a> ResolvedManifestPackages<'a> {
-    pub fn new(manifest_data: ChangedManifestPackages<'a>) -> Result<Self, BonjourError> {
-        let wapm_pkgs = manifest_data
-            .packages
-            .into_iter()
-            .filter_map(|k| match k {
-                PackageKey::WapmPackage(k) => Some(k),
-                _ => panic!("Non-wapm registry keys are not supported."),
-            })
-            .collect();
-        let packages: Vec<_> = Self::sync_packages(wapm_pkgs)
-            .map_err(|e| BonjourError::InstallError(e.to_string()))?;
-        Ok(Self { packages })
-    }
-
-    fn get_response(added_pkgs: Vec<WapmPackageKey<'a>>) -> get_packages_query::ResponseData {
-        let set: HashSet<WapmPackageKey<'a>> = added_pkgs.into_iter().collect();
-        let names = set.into_iter().map(|k| k.name.to_string()).collect();
-        let q = GetPackagesQuery::build_query(get_packages_query::Variables { names });
-        execute_query(&q).unwrap()
-    }
-
-    fn sync_packages(
-        added_packages: Vec<WapmPackageKey<'a>>,
-    ) -> Result<Vec<(WapmPackageKey<'a>, String)>, failure::Error> {
-        let response = Self::get_response(added_packages.clone());
-        let results: Vec<(WapmPackageKey<'a>, String)> = response
-            .package
-            .into_iter()
-            .filter_map(|p| p)
-            .filter_map(|p| {
-                let versions = p.versions.unwrap_or(vec![]);
-                let name = p.name;
-                Some((name, versions))
-            })
-            .map(|(n, vs)| {
-                vs.into_iter()
-                    .filter_map(|o| o)
-                    .map(|v| {
-                        let version = v.version;
-                        let download_url = v.distribution.download_url;
-                        (n.clone(), version, download_url)
-                    })
-                    .collect::<Vec<_>>()
-            })
-            .flatten()
-            .filter_map(|(n, v, d)| {
-                let key = added_packages.iter().find(|k| match k.name.find('/') {
-                    Some(_) => k.name == n && k.version == v,
-                    _ => k.name == &n[2..] && k.version == v,
-                });
-                key.map(|k| (k.clone(), d))
-            })
-            .collect();
-        Ok(results)
-    }
-}
-
-#[derive(Clone, Debug)]
-pub struct InstalledManifestPackages<'a> {
-    pub packages: Vec<(WapmPackageKey<'a>, Manifest, String)>,
-}
-
-impl<'a> InstalledManifestPackages<'a> {
-    pub fn install<P: AsRef<Path>>(
-        directory: P,
-        resolved_manifest_packages: ResolvedManifestPackages<'a>,
-    ) -> Result<Self, BonjourError> {
-        let packages_result: Result<Vec<(WapmPackageKey, PathBuf, String)>, BonjourError> =
-            resolved_manifest_packages
-                .packages
-                .into_iter()
-                .map(|(key, download_url)| Self::install_package(&directory, key, &download_url))
-                .collect();
-        let packages_result: Result<Vec<(WapmPackageKey, Manifest, String)>, BonjourError> =
-            packages_result?
-                .into_iter()
-                .map(|(key, dir, download_url)| {
-                    let m = Manifest::find_in_directory(&dir)
-                        .map(|m| (key, m))
-                        .map_err(|e| BonjourError::InstallError(e.to_string()));
-                    let m = m.map(|(k, m)| (k, m, download_url));
-                    m
-                })
-                .collect();
-        let packages = packages_result?;
-        Ok(Self { packages })
-    }
-
-    fn install_package<P: AsRef<Path>, S: AsRef<str>>(
-        directory: P,
-        key: WapmPackageKey<'a>,
-        download_url: S,
-    ) -> Result<(WapmPackageKey, PathBuf, String), BonjourError> {
-        let (namespace, pkg_name) = get_package_namespace_and_name(&key.name)
-            .map_err(|e| BonjourError::InstallError(e.to_string()))?;
-        let fully_qualified_package_name: String =
-            fully_qualified_package_display_name(pkg_name, &key.version);
-        let package_dir = create_package_dir(&directory, namespace, &fully_qualified_package_name)
-            .map_err(|_err| {
-                BonjourError::InstallError("Could not create package directory".to_string())
-            })?;
-        let mut response = reqwest::get(download_url.as_ref())
-            .map_err(|e| BonjourError::InstallError(e.to_string()))?;
-        let temp_dir = tempdir::TempDir::new("wapm_package_install").map_err(|_err| {
-            BonjourError::InstallError(
-                "Failed to create temporary directory to open the package in".to_string(),
-            )
-        })?;
-        let temp_tar_gz_path = temp_dir.path().join("package.tar.gz");
-        let mut dest = OpenOptions::new()
-            .read(true)
-            .write(true)
-            .create(true)
-            .open(&temp_tar_gz_path)
-            .map_err(|e| BonjourError::InstallError(e.to_string()))?;
-        io::copy(&mut response, &mut dest).map_err(|_err| {
-            BonjourError::InstallError("Could not copy response to temporary directory".to_string())
-        })?;
-        Self::decompress_and_extract_archive(dest, &package_dir)
-            .map_err(|err| BonjourError::InstallError(format!("{}", err)))?;
-        Ok((key, package_dir, download_url.as_ref().to_string()))
-    }
-
-    fn decompress_and_extract_archive<P: AsRef<Path>, F: io::Seek + io::Read>(
-        mut compressed_archive: F,
-        pkg_name: P,
-    ) -> Result<(), failure::Error> {
-        compressed_archive.seek(SeekFrom::Start(0))?;
-        let gz = GzDecoder::new(compressed_archive);
-        let mut archive = Archive::new(gz);
-        archive
-            .unpack(&pkg_name)
-            .map_err(|err| BonjourError::InstallError(format!("{}", err)))?;
-        Ok(())
-    }
-}
-
-#[derive(Clone, Debug)]
-pub struct ChangedManifestPackages<'a> {
-    pub packages: HashSet<PackageKey<'a>>,
-}
-
-impl<'a> ChangedManifestPackages<'a> {
-    pub fn prune_unchanged_dependencies(
-        manifest_data: ManifestData<'a>,
-        lockfile_data: &LockfileData<'a>,
-    ) -> Result<Self, BonjourError> {
-        let packages = match manifest_data.package_keys {
-            Some(m) => {
-                let lockfile_keys: HashSet<PackageKey<'a>> =
-                    lockfile_data.packages.keys().cloned().collect();
-                let differences: HashSet<PackageKey<'a>> =
-                    m.difference(&lockfile_keys).cloned().collect();
-                differences
-            }
-            _ => HashSet::new(),
-        };
-        Ok(Self { packages })
-    }
-}
-
 pub fn update<P: AsRef<Path>>(
     added_packages: Vec<(&str, &str)>,
     directory: P,
@@ -266,6 +74,8 @@ pub fn update<P: AsRef<Path>>(
     let manifest_lockfile_data =
         LockfileData::from_installed_packages(&installed_manifest_packages);
     let final_lockfile_data = manifest_lockfile_data.merge(lockfile_data);
+
+    manifest_result.update_manifest(added_packages)?;
     final_lockfile_data.generate_lockfile(&directory)?;
     Ok(())
 }
